refactor: exctract IssuanceAgent interface
This is a step towards adding an implementation using our public APIs.
